+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Implemented `/api/v2/notifications/policy` API endpoint.
 - Advertise RFC-9421 support using FEP-844e mechanism.
+- Implemented `/api/v1/statuses/{status_id}/favourited_by` API endpoint.
+- Implemented `/api/v1/pleroma/statuses/{status_id}/reactions` API endpoint.
 
 ### Changed